requirements_test.txt
-r requirements_dev.txt
pytest-homeassistant-custom-component
victron_ble @ git+https://github.com/notjj/victron-ble
bleak_retry_connector
serial
pyserial
aiousbwatcher
bluetooth_adapters
bluetooth_auto_recovery
bluetooth_data_tools
bluetooth_sensor_state_data
